+++ admin/setup.php
@@ -96,72 +96,94 @@ if (!class_exists('FormSetup')) {
 $formSetup = new FormSetup($db);
 $form = new Form($db);
 
-//
-//// Hôte
-//$item = $formSetup->newItem('NO_PARAM_JUST_TEXT');
-//$item->fieldOverride = (empty($_SERVER['HTTPS']) ? 'http://' : 'https://') . $_SERVER['HTTP_HOST'];
-//$item->cssClass = 'minwidth500';
-//
-//// Setup conf MYMODULE_MYPARAM1 as a simple string input
-//$item = $formSetup->newItem('MYMODULE_MYPARAM1');
-//$item->defaultFieldValue = 'default value';
-//
-//// Setup conf MYMODULE_MYPARAM1 as a simple textarea input but we replace the text of field title
-//$item = $formSetup->newItem('MYMODULE_MYPARAM2');
-//$item->nameText = $item->getNameText().' more html text ';
-//
-//// Setup conf MYMODULE_MYPARAM3
-//$item = $formSetup->newItem('MYMODULE_MYPARAM3');
-//$item->setAsThirdpartyType();
-//
-//// Setup conf MYMODULE_MYPARAM4 : exemple of quick define write style
-//$formSetup->newItem('MYMODULE_MYPARAM4')->setAsYesNo();
-//
-//// Setup conf MYMODULE_MYPARAM5
-//$formSetup->newItem('MYMODULE_MYPARAM5')->setAsEmailTemplate('thirdparty');
-//
-//// Setup conf MYMODULE_MYPARAM6
-//$formSetup->newItem('MYMODULE_MYPARAM6')->setAsSecureKey()->enabled = 0; // disabled
-//
-//// Setup conf MYMODULE_MYPARAM7
-////$formSetup->newItem('MYMODULE_MYPARAM7')->setAsProduct();
-//
-//$formSetup->newItem('Title')->setAsTitle();
-//
-//// Setup conf MYMODULE_MYPARAM8
-//$item = $formSetup->newItem('MYMODULE_MYPARAM8');
-//$TField = array(
-//	'test01' => $langs->trans('test01'),
-//	'test02' => $langs->trans('test02'),
-//	'test03' => $langs->trans('test03'),
-//	'test04' => $langs->trans('test04'),
-//	'test05' => $langs->trans('test05'),
-//	'test06' => $langs->trans('test06'),
-//);
-//$item->setAsMultiSelect($TField);
-//$item->helpText = $langs->transnoentities('MYMODULE_MYPARAM8');
-//
-//
-//// Setup conf MYMODULE_MYPARAM9
-//$formSetup->newItem('MYMODULE_MYPARAM9')->setAsSelect($TField);
-//
-//
-//// Setup conf MYMODULE_MYPARAM10
-//$item = $formSetup->newItem('MYMODULE_MYPARAM10');
-//$item->setAsColor();
-//$item->defaultFieldValue = '#FF0000';
-//$item->nameText = $item->getNameText().' more html text ';
-//$item->fieldInputOverride = '';
-//$item->helpText = $langs->transnoentities('AnHelpMessage');
-////$item->fieldValue = '';
-////$item->fieldAttr = array() ; // fields attribute only for compatible fields like input text
-////$item->fieldOverride = false; // set this var to override field output will override $fieldInputOverride and $fieldOutputOverride too
-////$item->fieldInputOverride = false; // set this var to override field input
-////$item->fieldOutputOverride = false; // set this var to override field output
-//
-
-
+/*
+ * DEMO
 // Hôte
+$item = $formSetup->newItem('NO_PARAM_JUST_TEXT');
+$item->fieldOverride = (empty($_SERVER['HTTPS']) ? 'http://' : 'https://') . $_SERVER['HTTP_HOST'];
+$item->cssClass = 'minwidth500';
+
+// Setup conf MYMODULE_MYPARAM1 as a simple string input
+$item = $formSetup->newItem('MYMODULE_MYPARAM1');
+$item->defaultFieldValue = 'default value';
+
+// Setup conf MYMODULE_MYPARAM1 as a simple textarea input but we replace the text of field title
+$item = $formSetup->newItem('MYMODULE_MYPARAM2');
+$item->nameText = $item->getNameText().' more html text ';
+
+// Setup conf MYMODULE_MYPARAM3
+$item = $formSetup->newItem('MYMODULE_MYPARAM3');
+$item->setAsThirdpartyType();
+
+// Setup conf MYMODULE_MYPARAM4 : exemple of quick define write style
+$formSetup->newItem('MYMODULE_MYPARAM4')->setAsYesNo();
+
+// Setup conf MYMODULE_MYPARAM5
+$formSetup->newItem('MYMODULE_MYPARAM5')->setAsEmailTemplate('thirdparty');
+
+// Setup conf MYMODULE_MYPARAM6
+$formSetup->newItem('MYMODULE_MYPARAM6')->setAsSecureKey()->enabled = 0; // disabled
+
+// Setup conf MYMODULE_MYPARAM7
+//$formSetup->newItem('MYMODULE_MYPARAM7')->setAsProduct();
+
+$formSetup->newItem('Title')->setAsTitle();
+
+
+
+
+
+
+// Setup conf MYMODULE_MYPARAM8
+$item = $formSetup->newItem('MYMODULE_MYPARAM8');
+$TField = array(
+	'test01' => $langs->trans('test01'),
+	'test02' => $langs->trans('test02'),
+	'test03' => $langs->trans('test03'),
+	'test04' => $langs->trans('test04'),
+	'test05' => $langs->trans('test05'),
+	'test06' => $langs->trans('test06'),
+);
+$item->setAsMultiSelect($TField);
+$item->helpText = $langs->transnoentities('MYMODULE_MYPARAM8');
+
+
+// Setup conf MYMODULE_MYPARAM9
+$formSetup->newItem('MYMODULE_MYPARAM9')->setAsSelect($TField);
+
+
+// Setup conf MYMODULE_MYPARAM10
+$item = $formSetup->newItem('MYMODULE_MYPARAM10');
+$item->setAsColor();
+$item->defaultFieldValue = '#FF0000';
+$item->nameText = $item->getNameText().' more html text ';
+$item->fieldInputOverride = '';
+$item->helpText = $langs->transnoentities('AnHelpMessage');
+//$item->fieldValue = '';
+//$item->fieldAttr = array() ; // fields attribute only for compatible fields like input text
+//$item->fieldOverride = false; // set this var to override field output will override $fieldInputOverride and $fieldOutputOverride too
+//$item->fieldInputOverride = false; // set this var to override field input
+//$item->fieldOutputOverride = false; // set this var to override field output
+*/
+
+
+$item = $formSetup->newItem('SP_MAX_SCRUM_TASK_STEP_QTY');
+$item->fieldAttr = array(
+	'type' => 'number',
+	'step' => '0.01',
+	'min' => 0
+);
+
+$item = $formSetup->newItem('SP_MAX_SCRUM_TASK_MAX_QTY');
+$item->fieldAttr = array(
+	'type' => 'number',
+	'step' => getDolGlobalString('SP_MAX_SCRUM_TASK_STEP_QTY'),
+	'min' => 0
+);
+
+$formSetup->htmlAfterOutputForm.='<script>$(document).on("change", "#setup-SP_MAX_SCRUM_TASK_STEP_QTY", function(){$("#setup-SP_MAX_SCRUM_TASK_MAX_QTY").attr("step", $(this).val());});</script>';
+
+
 $item = $formSetup->newItem('SCRUMPROJECT_DEFAULT_KANBAN_CONTACT_CODE');
 
 $sql = "SELECT code, libelle FROM  ".MAIN_DB_PREFIX."c_type_contact WHERE active=1 AND element='scrumproject_scrumcard' AND source='internal'";