+++ interface-liveupdate.php
@@ -84,7 +84,7 @@ function _actionLiveUpdate(&$jsonResponse){
 	// Todo use object display value like update form
 	$TAllowedObjects = array(
 		'scrumproject_scrumuserstorysprint' => array(
-			'allowedFields' => array('us_qty_planned')
+			'allowedFields' => array('us_qty_planned', 'label')
 		),
 		'scrumproject_scrumsprintuser' => array(
 			'allowedFields' => array('qty_availablity', 'availablity_rate', 'qty_velocity')