+++ ChangeLog.md
@@ -3,8 +3,23 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## Version 0.1 [2020-12-25]
 
+## Version 1.2
+
+- NEW : Planning interface for user stories *05/08/2022* - 1.2.0
+
+## Version 1.1
+
+- NEW : Split dialog for kanban card of type User story or Scrum task *28/07/2022* - 1.1.0
+
+## Version 1.0 [2020-12-25]
+
+- FIX : Suppression des champs qté réalisée et quantité prévue à la création d'un sprint *21/07/2022* - 1.0.9
+- FIX : Update visuel de la liste US planifiée *21/07/2022* - 1.0.8
+- FIX : SQL : Renvoi la liste de Tâches scrum de la fiche US planifiée concernée *21/07/2022* - 1.0.7
+- FIX : Fiche create et modify US planifiée et Tache Scrum : Suppression champ statut *21/07/2022* - 1.0.6
+- FIX : Formulaire U.S. et U.S. Planifiée type html champ description *21/07/2022* - 1.0.5
+- FIX : Modification langs Qty consommée *21/07/2022* - 1.0.4
 - FIX : scrumuserstorysprint status BTN  *03/05/2022* - 1.0.3
 - FIX : sql query for scrum quntity summary  *03/05/2022* - 1.0.2
 - FIX : elements names *03/05/2022* - 1.0.1